Move `Resources` into it's own module
