+++ agape/Cargo.toml
@@ -17,7 +17,6 @@ fontdue = "0.9.3"
 agape_core = {version = "0.1.0",path = "../agape_core"}
 agape_macros = {version = "0.1.0",path = "../agape_macros"}
 agape_layout = {version = "0.1.0",path = "../agape_layout"}
-resvg = "0.44.0"
 reqwest = {version = "0.12.12",features = ["blocking","stream","json"]}
 tiny-skia.workspace = true
 pixels.workspace = true
