+++ agape/src/system.rs
@@ -1,4 +1,4 @@
-//! Systems are functions that are stored in the [`App`] and run every frame.
+//! Systems are stored procedures that run every frame.
 //! They have a `&mut` to the global [`Resources`] allowing it to be modified.
 //!
 //! # Create a system
@@ -15,7 +15,7 @@
 
 use agape_core::{Position, Size};
 use std::any::Any;
-use agape_layout::Layout;
+use crate::Resources;
 
 /// A [`System`] is a stored procedure.
 pub trait System {
@@ -53,63 +53,3 @@ where
         (self.f)(resources)
     }
 }
-
-/// Global resources
-pub struct Resources {
-    items: Vec<Box<dyn Any>>,
-}
-
-impl Resources {
-    pub fn new() -> Resources {
-        Self { items: vec![] }
-    }
-
-    /// Insert a resource.
-    pub fn insert<T: 'static>(&mut self, item: T) {
-        // Don't insert the same resource twice
-        if let None = self.get::<T>() {
-            self.items.push(Box::new(item));
-        }
-    }
-
-    /// Retrieve a resource of type `T`.
-    pub fn get<T: 'static>(&self) -> Option<&T> {
-        for item in &self.items {
-            match item.downcast_ref::<T>() {
-                Some(item) => return Some(item),
-                None => continue,
-            }
-        }
-
-        None
-    }
-    
-    /// Retrieve an owned resource of type `T`.
-    pub fn get_owned<T: Clone + 'static>(&self) -> Option<T> {
-        match self.get::<T>() { 
-            Some(item) => Some(item.clone()),
-            None => None,
-        }
-    }
-
-    /// Retrieve a mutable reference to a resource of type `T`.
-    pub fn get_mut<T: 'static>(&mut self) -> Option<&mut T> {
-        for items in &mut self.items {
-            match items.downcast_mut::<T>() {
-                Some(item) => return Some(item),
-                None => continue,
-            }
-        }
-
-        None
-    }
-}
-
-/// The current cursor position.
-#[derive(Debug, Default,Copy, Clone)]
-pub struct CursorPosition(pub Position);
-
-/// The window size.
-#[derive(Debug, Default,Copy, Clone)]
-pub struct WindowSize(pub Size);
-
